"Completion" of the uploadX file
Visualisation of match data in a table and sorting by match number is complete using the 2017 JSON object format. Changes to Tele.js and SplashSheet.html aimed to ensure that new naming schemes for the passing of data from splash sheet to tele are functional and that it still works as intended with the new names.

uploadX does not actually upload data, as it lacks the upload function, but does sort data and react visually as desired.